Knubot scripting
/!\ Recreate your database /!\

To enable a knubot script, proceed like this:
Spawn and save a npc (like Ofoz, hash PUEU)
then edit mobspawns table and enter the method name of the script (i.e. InitializeFlappy as in the KnuBotFlappy.cs script)
This script then will be called on spawning of this particular mob.
For more explanation how to build a KnuBot Tree, refer to KnuBotFlappy in the Scripts folder or ask me in irc